Only composite avatars in timelines, not elsewhere (e.g., the account page)
